extract cobra commands into functions
